feat: Lapnoteページのコンポーネントを分割し、タイマーと書籍情報表示を実装
